refactor: extract shared relativeTime() helper from Feed/Ledger/Jobs
The relativeTime() function was copy-pasted identically across
FeedSection.tsx, LedgerSection.tsx, and JobsSection.tsx, with a
TODO already in JobsSection.tsx calling for this extraction.

- Create app/src/agentworld/utils/relativeTime.ts with the canonical
  implementation
- Replace all three local copies with import from the shared util
- Remove the now-resolved TODO comment

Closes #4427
