+++ app/src/agentworld/pages/FeedSection.tsx
@@ -32,6 +32,7 @@ import {
 import { fetchWalletStatus } from '../../services/walletApi';
 import { apiClient } from '../AgentWorldShell';
 import ConfirmDialog from '../components/ConfirmDialog';
+import { relativeTime } from '../utils/relativeTime';
 
 const log = debug('agentworld:feed');
 
@@ -64,17 +65,6 @@ type WalletConfigured = 'resolving' | 'no' | 'yes' | 'unknown';
 type WalletResolution = { agentId: string | null; configured: WalletConfigured };
 
 // ── Helpers ───────────────────────────────────────────────────────────────────
-
-function relativeTime(iso: string): string {
-  const ms = Date.now() - new Date(iso).getTime();
-  const mins = Math.floor(ms / 60000);
-  if (mins < 1) return 'just now';
-  if (mins < 60) return `${mins}m ago`;
-  const hrs = Math.floor(mins / 60);
-  if (hrs < 24) return `${hrs}h ago`;
-  const days = Math.floor(hrs / 24);
-  return `${days}d ago`;
-}
 
 function isWalletLocked(message: string): boolean {
   return (
